tune: available_backends takes dtype; no flyc on fp32
flyc is f16/bf16 WMMA only (aot/_flyc_common.py:64-66, flyc_attn_fwd.py:143
disable the functional on '*fp32' in Q). Left in the sweep it would not just
waste a run: flyc refuses with hipErrorNotSupported, which kftdesc's
check_early_reject_results does not recognise -- it knows only
hipErrorPeerAccessUnsupported -- so every fp32 entry would land in the database
as a failure rather than an honest non-answer. fp32 is on by default.

enumerate_variants passes entry.dtype; desc.py passes im.dtype at both
create_extargs call sites, where im is FlashInputMetadata(**asdict(entry)), so
both filters see the same value and positions stay in sync. level_kernel.py and
reference.py accept dtype and ignore it -- desc.py reaches every level through
one call site.

Verified over arch x dtype x iface with pyaotriton stubbed: flyc is absent for
float32 on gfx950/gfx1201 and present for f16/bf16; every position still
resolves to the force_backend_index its name implies. python/test: 309 passed,
1 skipped.

+++ modules/flash/tune/desc.py
@@ -283,7 +283,7 @@ class FlashTune(TuningDescription):
         from aotriton.tune.gpu_utils import device_ctx, default_device_string
         with device_ctx():
             kernel = self.get_impl(which_impl.dsl_name)
-            args = kernel.create_extargs(which_impl=which_impl)
+            args = kernel.create_extargs(which_impl=which_impl, dtype=im.dtype)
             d = torch.load(pt, map_location=default_device_string(), mmap=True)
             inputs = from_dict(data_class=kernel.PT_INPUT_CLASS, data=d["bidi_inputs"], config=dacite_tuple)
             direct_inputs = kernel.prepare_directs(im, inputs)
@@ -309,7 +309,7 @@ class FlashTune(TuningDescription):
         from aotriton.tune.gpu_utils import do_bench, device_ctx, default_device_string
         with device_ctx():
             kernel = self.get_impl(which_impl.dsl_name)
-            args = kernel.create_extargs(which_impl=which_impl, probe=True)
+            args = kernel.create_extargs(which_impl=which_impl, probe=True, dtype=im.dtype)
             d = torch.load(pt, map_location=default_device_string(), mmap=True)
             inputs = from_dict(data_class=kernel.PT_INPUT_CLASS, data=d["bidi_inputs"], config=dacite_tuple)
             direct_inputs = kernel.prepare_directs(im, inputs)
